pyproject.toml
[project]
name = "meco"
version = "0.1.0"
description = "Micro-evolutionary code optimizer using LLM-generated variants and forward selection."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.66.5,<2.0.0",
    "pydantic>=2.10.6,<3.0.0",
    "docker>=7.1.0,<8.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
